+++ molecool/__init__.py
@@ -4,7 +4,11 @@ A Python package for analyzing and visualizing xyz files. For MolSSI Workshop Py
 """
 
 # Add imports here
-from .functions import *
+from .measure import calculate_distance, calculate_angle
+from .molecule import build_bond_list
+from .visualize import draw_molecule, bond_histogram
+from . import io
+
 
 # Handle versioneer
 from ._version import get_versions
