+++ molecool/data/pdb/benzene.pdb
@@ -1,0 +1,31 @@
+HETATM    1  C01 UNK     0      -0.010   0.000   0.000  1.00  0.00           C
+HETATM    2  C02 UNK     0      -0.707   1.208   0.000  1.00  0.00           C
+HETATM    3  C03 UNK     0      -2.102   1.208   0.000  1.00  0.00           C
+HETATM    4  C04 UNK     0      -2.799   0.000   0.000  1.00  0.00           C
+HETATM    5  C05 UNK     0      -2.102  -1.208   0.000  1.00  0.00           C
+HETATM    6  C06 UNK     0      -0.707  -1.208   0.000  1.00  0.00           C
+HETATM    7  H07 UNK     0       1.093   0.000   0.000  1.00  0.00           H
+HETATM    8  H08 UNK     0      -0.156   2.163   0.000  1.00  0.00           H
+HETATM    9  H09 UNK     0      -2.653   2.163   0.000  1.00  0.00           H
+HETATM   10  H10 UNK     0      -3.902   0.000   0.000  1.00  0.00           H
+HETATM   11  H11 UNK     0      -2.653  -2.163   0.000  1.00  0.00           H
+HETATM   12  H12 UNK     0      -0.156  -2.163   0.000  1.00  0.00           H
+CONECT    1    6    7    2
+CONECT    1    2
+CONECT    2    1    3    8
+CONECT    2    1
+CONECT    3    2    9    4
+CONECT    3    4
+CONECT    4    3    5   10
+CONECT    4    3
+CONECT    5    4   11    6
+CONECT    5    6
+CONECT    6    1    5   12
+CONECT    6    5
+CONECT    7    1
+CONECT    8    2
+CONECT    9    3
+CONECT   10    4
+CONECT   11    5
+CONECT   12    6
+END
