+++ molecool/data/pdb/water.pdb
@@ -1,0 +1,5 @@
+CRYST1    0.000    0.000    0.000  90.00  90.00  90.00 P 1           1
+ATOM      1  H   HOH X   1       9.626   6.787  12.673  0.00  0.00           H
+ATOM      2  H   HOH X   1       9.626   8.420  12.673  0.00  0.00           H
+ATOM      3  O   HOH X   1      10.203   7.604  12.673  0.00  0.00           O            
+END
