+++ molecool/molecule.py
@@ -1,0 +1,78 @@
+"""
+molecule.py
+A Python package for analyzing and visualizing xyz files. For MolSSI Workshop Python Package development workshop.
+
+Handles the primary functions
+"""
+
+from .measure import calculate_distance
+
+def build_bond_list(coordinates, max_bond=1.5, min_bond=0):
+
+    # Find the bonds in a molecule (set of coordinates) based on distance criteria.
+    bonds = {}
+    num_atoms = len(coordinates)
+
+    for atom1 in range(num_atoms):
+        for atom2 in range(atom1, num_atoms):
+            distance = calculate_distance(coordinates[atom1], coordinates[atom2])
+            if distance > min_bond and distance < max_bond:
+                bonds[(atom1, atom2)] = distance
+
+    return bonds
+
+
+
+def canvas(with_attribution=True):
+    """
+    Placeholder function to show example docstring (NumPy format)
+
+    Replace this function and doc string for your own project
+
+    Parameters
+    ----------
+    with_attribution : bool, Optional, default: True
+        Set whether or not to display who the quote is from
+
+    Returns
+    -------
+    quote : str
+        Compiled string including quote and optional attribution
+    """
+
+    quote = "The code is but a canvas to our imagination."
+    if with_attribution:
+        quote += "\n\t- Adapted from Henry David Thoreau"
+    return quote
+
+
+def zen(with_attribution=True):
+    quote = """Beautiful is better than ugly.
+    Explicit is better than implicit.
+    Simple is better than complex.
+    Complex is better than complicated.
+    Flat is better than nested.
+    Sparse is better than dense.
+    Readability counts.
+    Special cases aren't special enough to break the rules.
+    Although practicality beats purity.
+    Errors should never pass silently.
+    Unless explicitly silenced.
+    In the face of ambiguity, refuse the temptation to guess.
+    There should be one-- and preferably only one --obvious way to do it.
+    Although that way may not be obvious at first unless you're Dutch.
+    Now is better than never.
+    Although never is often better than *right* now.
+    If the implementation is hard to explain, it's a bad idea.
+    If the implementation is easy to explain, it may be a good idea.
+    Namespaces are one honking great idea -- let's do more of those!"""
+
+    if with_attribution:
+        quote += "\n\tTim Peters"
+
+    return quote
+
+
+if __name__ == "__main__":
+    # Do something if this file is invoked on its own
+    print(canvas())
